requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
pillow
scipy
flask
